🐛 fix: likedPosts,Comments 에러 수정

+++ src/lib/post/api.ts
@@ -91,9 +91,57 @@ export const updatePost = async (postId: string, postDto: postDto) => {
   });
 };
 
+// 모든 사용자의 like_posts 컬렉션에서 특정 포스트 ID의 좋아요 정보를 삭제하는 함수
+const deleteAllPostLikes = async (postId: string) => {
+  try {
+    // 모든 사용자 문서를 가져옴
+    const usersSnapshot = await getDocs(collection(db, "users"));
+
+    if (usersSnapshot.empty) {
+      console.log("No users found.");
+      return;
+    }
+
+    const deletePromises: Promise<void>[] = [];
+
+    for (const userDoc of usersSnapshot.docs) {
+      const userId = userDoc.id;
+      const likePostsQuery = query(
+        collection(db, `like_posts/${userId}/posts`),
+        where("postId", "==", postId)
+      );
+      const likePostsSnapshot = await getDocs(likePostsQuery);
+
+      likePostsSnapshot.forEach((likeDoc) => {
+        console.log("Deleting likeDoc with path:", likeDoc.ref.path);
+        deletePromises.push(deleteDoc(likeDoc.ref));
+      });
+    }
+
+    await Promise.all(deletePromises);
+    console.log("All like documents deleted successfully.");
+  } catch (error) {
+    console.error("Failed to delete post likes: ", error);
+    throw new Error("포스트 좋아요 정보를 삭제하는 데 실패하였습니다.");
+  }
+};
+
 export const deletePost = async (postId: string) => {
-  const postRef = doc(db, "posts", postId);
-  await deleteDoc(postRef);
+  try {
+    const postRef = doc(db, "posts", postId);
+    const postSnap = await getDoc(postRef);
+    if (!postSnap.exists()) {
+      throw new Error("포스트를 찾을 수 없습니다.");
+    }
+
+    // 포스트에 대한 모든 좋아요 정보 삭제
+    await deleteAllPostLikes(postId);
+
+    // 포스트 삭제
+    await deleteDoc(postRef);
+  } catch (error) {
+    throw new Error("포스트 삭제에 실패하였습니다.");
+  }
 };
 
 export const getAllPosts = async (
@@ -123,8 +171,6 @@ export const getPostByPostId = async (postId: string) => {
   if (docSnap.exists()) {
     const postData = docSnap.data();
     return postData;
-  } else {
-    throw new Error("포스트를 찾을 수 없습니다.");
   }
 };
